mingw casting.. int shoulld be enough

+++ gstreamer-encoder.c
@@ -112,27 +112,27 @@ void *gstreamer_encoder_create(obs_data_t *settings, obs_encoder_t *encoder)
 	gchar *encoder_string = "";
 	if (g_strcmp0(encoder_type, "x264") == 0) {
 		encoder_string = g_strdup_printf(
-			"x264enc tune=zerolatency bitrate=%lld key-int-max=%lld",
-			obs_data_get_int(data->settings, "bitrate"),
-			obs_data_get_int(data->settings, "keyint_sec") *
+			"x264enc tune=zerolatency bitrate=%d key-int-max=%d",
+			(int)obs_data_get_int(data->settings, "bitrate"),
+			(int)obs_data_get_int(data->settings, "keyint_sec") *
 				data->ovi.fps_num / data->ovi.fps_den);
 	} else if (g_strcmp0(encoder_type, "nvh264enc") == 0) {
 		encoder_string = g_strdup_printf(
-			"nvh264enc bitrate=%lld gop-size=%lld",
-			obs_data_get_int(data->settings, "bitrate"),
-			obs_data_get_int(data->settings, "keyint_sec") *
+			"nvh264enc bitrate=%d gop-size=%d",
+			(int)obs_data_get_int(data->settings, "bitrate"),
+			(int)obs_data_get_int(data->settings, "keyint_sec") *
 				data->ovi.fps_num / data->ovi.fps_den);
 	} else if (g_strcmp0(encoder_type, "vaapih264enc") == 0) {
 		encoder_string = g_strdup_printf(
-			"vaapih264enc bitrate=%lld keyframe-period=%lld",
-			obs_data_get_int(data->settings, "bitrate"),
-			obs_data_get_int(data->settings, "keyint_sec") *
+			"vaapih264enc bitrate=%d keyframe-period=%d",
+			(int)obs_data_get_int(data->settings, "bitrate"),
+			(int)obs_data_get_int(data->settings, "keyint_sec") *
 				data->ovi.fps_num / data->ovi.fps_den);
 	} else if (g_strcmp0(encoder_type, "omxh264enc") == 0) {
 		encoder_string = g_strdup_printf(
-			"omxh264enc target-bitrate=%lld periodicty-idr=%lld",
-			obs_data_get_int(data->settings, "bitrate") * 1000,
-			obs_data_get_int(data->settings, "keyint_sec") *
+			"omxh264enc target-bitrate=%d periodicty-idr=%d",
+			(int)obs_data_get_int(data->settings, "bitrate") * 1000,
+			(int)obs_data_get_int(data->settings, "keyint_sec") *
 				data->ovi.fps_num / data->ovi.fps_den);
 	} else if (g_strcmp0(encoder_type, "vtenc_h264") == 0) {
 		encoder_string = g_strdup_printf(
